moved cache test in the correct place
