+++ macros/src/test/scala/gestalt/macros/AnnotationMacroTest.scala
@@ -23,6 +23,21 @@ class AnnotationMacroTest extends TestSuite {
     val note = new Note("Vassily", "Pupkin", "hello", "This is a test!")
   }
 
+  test("cache") {
+    import scala.util.Random
+
+    @cache
+    def rand(seed: Int): Random = new Random(seed)
+
+    assert(_rand == null)
+    val rand1 = rand(3)
+    assert(rand1 != null)
+    assert(_rand != null)
+
+    val rand2 = rand(4)
+    assert(rand1 == rand2)
+  }
+
   // need to put at top-level for indexing to make them available for @dynamic
   case class Authorized(val i: Int)
   case class Token(val i: Int)
